tests/wpt/web-platform-tests/offscreen-canvas/line-styles/2d.line.defaults.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.line.defaults</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.line.defaults</h1>
<p class="desc"></p>


<script>
var t = async_test("");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

_assertSame(ctx.lineWidth, 1, "ctx.lineWidth", "1");
_assertSame(ctx.lineCap, 'butt', "ctx.lineCap", "'butt'");
_assertSame(ctx.lineJoin, 'miter', "ctx.lineJoin", "'miter'");
_assertSame(ctx.miterLimit, 10, "ctx.miterLimit", "10");

t.done();

});
</script>
